unify access methods into current_user
