+++ app/controllers/builds_controller.rb
@@ -1,5 +1,5 @@
 class BuildsController < ApplicationController
-  include ProjectLevelAuthorization
+  include CurrentProject
 
   before_action :authorize_project_deployer!
