+++ app/controllers/commit_statuses_controller.rb
@@ -1,5 +1,5 @@
 class CommitStatusesController < ApplicationController
-  include ProjectLevelAuthorization
+  include CurrentProject
 
   before_action :authorize_project_deployer!
