+++ app/controllers/concerns/current_user.rb
@@ -48,7 +48,15 @@ module CurrentUser
     unauthorized! unless current_user.is_admin?
   end
 
+  def authorize_project_admin!
+    unauthorized! unless current_user.is_admin_for?(current_project)
+  end
+
   def authorize_deployer!
     unauthorized! unless current_user.is_deployer?
+  end
+
+  def authorize_project_deployer!
+    unauthorized! unless current_user.is_deployer_for?(current_project)
   end
 end
