+++ app/controllers/macros_controller.rb
@@ -1,5 +1,5 @@
 class MacrosController < ApplicationController
-  include ProjectLevelAuthorization
+  include CurrentProject
 
   before_action :authorize_project_deployer!
   before_action :authorize_project_admin!, only: [:new, :create, :edit, :update]
