+++ app/controllers/releases_controller.rb
@@ -1,5 +1,5 @@
 class ReleasesController < ApplicationController
-  include ProjectLevelAuthorization
+  include CurrentProject
 
   before_action :authorize_project_deployer!, except: [:show, :index]
