+++ plugins/kubernetes/app/controllers/kubernetes_project_controller.rb
@@ -1,5 +1,5 @@
 class KubernetesProjectController < ApplicationController
-  include ProjectLevelAuthorization
+  include CurrentProject
   before_action :authorize_project_deployer!
 
   def show
